chore: remove unused import and clean up
