+++ .env.defaults
@@ -35,3 +35,5 @@ MONGO_URL="mongodb://localhost:27017"
 SPOTIFY="https://api.spotify.com/v1"
 SPOTIFY_ID="SPOTIFY_ID"
 SPOTIFY_SECRET="SPOTIFY_SECRET"
+
+PORT=8080
