+++ src/config/local/index.ts
@@ -1,19 +1,1 @@
-import { Collection } from 'x/mongo';
-import { logger } from '../../common/core/logger.ts';
-import { ConfigDocument } from './types.d.ts';
-
-export class LocalSource {
-  constructor(
-    private readonly collection?: Collection<ConfigDocument>,
-  ) {}
-
-  getConfig = async (): Promise<ConfigDocument | undefined> => {
-    const config = await this.collection?.findOne()
-      ?.catch((e) => {
-        logger.error(`Unable to find config in collection`, e);
-        return undefined;
-      });
-
-    return config;
-  };
-}
+export * from './source.ts';
