+++ src/news/local/index.ts
@@ -1,6 +1,5 @@
 import { Collection } from 'x/mongo';
 import { logger } from '../../common/core/logger.ts';
-import { pagination } from '../../common/core/utils.ts';
 import { IPaging } from '../../common/types/paging.d.ts';
 import { IResponse } from '../../common/types/response.d.ts';
 import { fromEntity, toEntity } from '../mapper.ts';
