+++ src/series/local/transformer.ts
@@ -2,9 +2,15 @@ import { Document } from 'x/mongo';
 import { Transform } from '../../common/transformer/types.d.ts';
 import { MediaWithSeason } from '../types.d.ts';
 
-export const transform: Transform<Document, MediaWithSeason> = (
+export const transform: Transform<
+  Document | undefined,
+  MediaWithSeason | undefined
+> = (
   sourceData,
 ) => {
-  const { _id, ...rest } = sourceData;
-  return rest as MediaWithSeason;
+  if (sourceData) {
+    const { _id, ...rest } = sourceData;
+    return rest as MediaWithSeason;
+  }
+  return undefined;
 };
